feat: Finish testing main service start merging two services

+++ ewm-service/src/main/java/ru/practicum/category/service/CategoryServiceImpl.java
@@ -6,9 +6,14 @@ import ru.practicum.category.dto.CategoryDto;
 import ru.practicum.category.mapper.CategoryMapper;
 import ru.practicum.category.model.Category;
 import ru.practicum.category.repository.CategoryRepository;
+import ru.practicum.event.model.Event;
+import ru.practicum.event.repository.EventRepository;
+import ru.practicum.exception.CategoryNameOccupiedException;
+import ru.practicum.exception.CategoryRelatedEventsException;
 import ru.practicum.exception.EntityNotFoundException;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -18,10 +23,14 @@ public class CategoryServiceImpl implements CategoryService {
 
 
     private final CategoryRepository repository;
+    private final EventRepository eventRepository;
     private final CategoryMapper mapper;
 
     @Override
     public CategoryDto add(CategoryDto categoryDto) {
+        Optional<Category> optionalCategory = repository.findByName(categoryDto.getName());
+        if (optionalCategory.isPresent())
+            throw new CategoryNameOccupiedException(categoryDto.getName());
         return mapper.mapToCategoryDto(repository.save(mapper.mapToCategory(categoryDto)));
     }
 
@@ -41,11 +50,17 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public void remove(int catId) {
         Category category = mapper.mapToCategory(getById(catId));
+        List<Event> events = eventRepository.findByCategory_Id(catId);
+        if (events.size() > 0)
+            throw new CategoryRelatedEventsException(catId);
         repository.delete(category);
     }
 
     @Override
     public CategoryDto update(int catId, CategoryDto category) {
+        Optional<Category> optionalCategory = repository.findByName(category.getName());
+        if (optionalCategory.isPresent() && optionalCategory.get().getId() != catId)
+            throw new CategoryNameOccupiedException(category.getName());
         Category oldCategory = mapper.mapToCategory(getById(catId));
         return mapper.mapToCategoryDto(repository.save(mapper.mapToCategory(category)));
     }
